+++ docs/meta/history.md
@@ -14,6 +14,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 This is the first release candidate for Hatch v1, a complete rewrite.
 
+-----
+
 ## **Hatchling**
 
 ### Unreleased