+++ docs/meta/history.md
@@ -2,7 +2,11 @@
 
 -----
 
-## Hatch
+All notable changes to this project will be documented in this file.
+
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## **Hatch**
 
 ### Unreleased
 
@@ -10,7 +14,9 @@
 
 This is the first release candidate for Hatch v1, a complete rewrite.
 
-## Hatchling
+-----
+
+## **Hatchling**
 
 ### Unreleased
 
@@ -21,6 +27,7 @@ This is the first release candidate for Hatch v1, a complete rewrite.
 ***Added:***
 
 - Refactor builder config handling into its own class
+- Allow build hooks to access builder configuration
 
 ### [0.3.1](https://github.com/ofek/hatch/releases/tag/hatchling-v0.3.1) - 2021-12-30 ### {: #hatchling-v0.3.1 }